Move header files to the include directory
Developers seeking to use the API only need the header files and a
compiled version of the DLL.
