+++ include/game_patch.h
