+++ include/game_patch/game_buffer_patch.h
