+++ include/game_patch/game_interception_patch.h
