+++ include/game_patch/game_patch_base.h
